+++ App/Viewer/DlgSettings.h
@@ -20,6 +20,10 @@ private slots:
     void on_pbOk_clicked();
     void on_pbNo_clicked();
     
+    void on_cbEnableViewPassword_clicked(bool checked);
+    
+    void on_pbEncryptKey_clicked();
+    
 private:
     Ui::CDlgSettings *ui;