+++ Plugins/LibVNCServer/Viewer/DlgSettingsLibVnc.cpp
@@ -3,6 +3,7 @@
 #include "DlgSettingsLibVnc.h"
 #include "ui_DlgSettingsLibVnc.h"
 #include <QDebug>
+#include "ManagePassword.h"
 
 CDlgSettingsLibVnc::CDlgSettingsLibVnc(CConnecterLibVNCServer *pConnecter, QWidget *parent) :
     QDialog(parent),
@@ -18,6 +19,7 @@ CDlgSettingsLibVnc::CDlgSettingsLibVnc(CConnecterLibVNCServer *pConnecter, QWidg
     ui->spPort->setValue(m_pPara->GetPort());
     ui->leUserName->setText(m_pPara->GetUser());
     ui->lePassword->setText(m_pPara->GetPassword());
+    ui->pbShow->setEnabled(CManagePassword::Instance()->GetViewPassowrd());
     ui->cbSavePassword->setChecked(m_pPara->GetSavePassword());
     
     ui->cbShared->setChecked(m_pPara->GetShared());