+++ Plugins/TigerVnc/Viewer/DlgSettingsTigerVnc.cpp
@@ -5,6 +5,7 @@
 #include "rfb/encodings.h"
 #include <QDebug>
 #include <QMessageBox>
+#include "ManagePassword.h"
 
 #ifdef HAVE_QXMPP
     #include "QXmppUtils.h"
@@ -223,6 +224,7 @@ void CDlgSettingsTigerVnc::showEvent(QShowEvent *event)
     ui->leName->setText(m_pPara->GetName());
     ui->leUserName->setText(m_pPara->GetUser());
     ui->lePassword->setText(m_pPara->GetPassword());
+    ui->pbShow->setEnabled(CManagePassword::Instance()->GetViewPassowrd());
     ui->cbSave->setChecked(m_pPara->GetSavePassword());
     ui->cbOnlyView->setChecked(m_pPara->GetOnlyView());